bug: rename custom icons
